BuildConfigsの切り出し
